Add rule even if expired rule uses that port
Fixes #7 by adding an additional check when a new rule matches an
existing rule's port to see if the existing rule has expired. A warning
is shown to the user, but they may proceed if they want to.

Also, added an improved explanation of the rules that NATMGR uses to the
README.