Listing rules now does not require root privs
By default, when an instance of the `Manager` class is created, it will
check if the program is running with root privileges. Commands that
don't require this (e.g. `list`) can explicitly disable this. Any
actions that require root (such as changing the configuration file)
still obtain it before executing, this is mostly just a convenience for
`list`.

Fixes #2